Allow attaching service pods to additional networks
For network isolation outgoing traffic, service pods need to be
attached additional networks. To achive this, the the
`k8s.v1.cni.cncf.io/networks` annotation needs to be set on the
pods. This adds a networkAttachments parameter to the
CRD where it can be specified to which additional networks the
service pods should be attached to.

Depends-On: openstack-k8s-operators/lib-common#133